Add folder hierarchy
