+++ src/core/ResourceGuesser.test.tsx
