+++ src/edit/EditGuesser.tsx
@@ -1,62 +1,67 @@
 import React from 'react';
 import {
-  Create,
+  Edit,
   FormTab,
   SimpleForm,
   TabbedForm,
   useResourceContext,
 } from 'react-admin';
+import { useParams } from 'react-router-dom';
 import type { Field, Resource } from '@api-platform/api-doc-parser';
 
-import InputGuesser from './InputGuesser.js';
-import Introspecter from './Introspecter.js';
-import useDisplayOverrideCode from './useDisplayOverrideCode.js';
-import useOnSubmit from './useOnSubmit.js';
+import InputGuesser from '../input/InputGuesser.js';
+import Introspecter from '../introspection/Introspecter.js';
+import useMercureSubscription from '../mercure/useMercureSubscription.js';
+import useDisplayOverrideCode from '../useDisplayOverrideCode.js';
+import useOnSubmit from '../useOnSubmit.js';
 import type {
-  CreateGuesserProps,
-  IntrospectedCreateGuesserProps,
-} from './types.js';
+  EditGuesserProps,
+  IntrospectedEditGuesserProps,
+} from '../types.js';
 
 const getOverrideCode = (schema: Resource, fields: Field[]) => {
   let code =
-    'If you want to override at least one input, paste this content in the <CreateGuesser> component of your resource:\n\n';
+    'If you want to override at least one input, paste this content in the <EditGuesser> component of your resource:\n\n';
 
-  code += `const ${schema.title}Create = props => (\n`;
-  code += `    <CreateGuesser {...props}>\n`;
+  code += `const ${schema.title}Edit = props => (\n`;
+  code += `    <EditGuesser {...props}>\n`;
 
   fields.forEach((field) => {
     code += `        <InputGuesser source={"${field.name}"} />\n`;
   });
-  code += `    </CreateGuesser>\n`;
+  code += `    </EditGuesser>\n`;
   code += `);\n`;
   code += `\n`;
   code += `And don't forget update your <ResourceGuesser> component:\n`;
-  code += `<ResourceGuesser name={"${schema.name}"} create={${schema.title}Create} />`;
+  code += `<ResourceGuesser name={"${schema.name}"} edit={${schema.title}Edit} />`;
 
   return code;
 };
 
-export const IntrospectedCreateGuesser = ({
+export const IntrospectedEditGuesser = ({
   fields,
   readableFields,
   writableFields,
   schema,
   schemaAnalyzer,
   resource,
+  mutationMode = 'pessimistic',
   mutationOptions,
   redirect: redirectTo = 'list',
   mode,
   defaultValues,
-  transform,
   validate,
+  transform,
   toolbar,
   warnWhenUnsavedChanges,
-  sanitizeEmptyValues = true,
   formComponent,
   viewComponent,
+  sanitizeEmptyValues = true,
   children,
   ...props
-}: IntrospectedCreateGuesserProps) => {
+}: IntrospectedEditGuesserProps) => {
+  const { id: routeId } = useParams<'id'>();
+  const id = decodeURIComponent(routeId ?? '');
   const save = useOnSubmit({
     resource,
     schemaAnalyzer,
@@ -65,6 +70,7 @@ export const IntrospectedCreateGuesser = ({
     transform,
     redirectTo,
   });
+  useMercureSubscription(resource, id);
 
   const displayOverrideCode = useDisplayOverrideCode();
 
@@ -83,9 +89,15 @@ export const IntrospectedCreateGuesser = ({
   const FormType = hasFormTab ? TabbedForm : SimpleForm;
 
   return (
-    <Create resource={resource} component={viewComponent} {...props}>
+    <Edit
+      resource={resource}
+      id={id}
+      mutationMode={mutationMode}
+      redirect={redirectTo}
+      component={viewComponent}
+      {...props}>
       <FormType
-        onSubmit={save}
+        onSubmit={mutationMode !== 'pessimistic' ? undefined : save}
         mode={mode}
         defaultValues={defaultValues}
         validate={validate}
@@ -95,11 +107,11 @@ export const IntrospectedCreateGuesser = ({
         component={formComponent}>
         {inputChildren}
       </FormType>
-    </Create>
+    </Edit>
   );
 };
 
-const CreateGuesser = (props: CreateGuesserProps) => {
+const EditGuesser = (props: EditGuesserProps) => {
   const resource = useResourceContext(props);
   if (!resource) {
     throw new Error('guesser must be used with a resource');
@@ -107,11 +119,11 @@ const CreateGuesser = (props: CreateGuesserProps) => {
 
   return (
     <Introspecter
-      component={IntrospectedCreateGuesser}
+      component={IntrospectedEditGuesser}
       resource={resource}
       {...props}
     />
   );
 };
 
-export default CreateGuesser;
+export default EditGuesser;
