+++ src/field/FieldGuesser.test.tsx
@@ -4,14 +4,14 @@ import { Resource } from '@api-platform/api-doc-parser';
 import { render, screen, waitFor } from '@testing-library/react';
 
 import FieldGuesser from './FieldGuesser.js';
-import SchemaAnalyzerContext from './SchemaAnalyzerContext.js';
-import schemaAnalyzer from './hydra/schemaAnalyzer.js';
+import SchemaAnalyzerContext from '../introspection/SchemaAnalyzerContext.js';
+import schemaAnalyzer from '../hydra/schemaAnalyzer.js';
 import type {
   ApiPlatformAdminDataProvider,
   ApiPlatformAdminRecord,
-} from './types.js';
+} from '../types.js';
 
-import { API_FIELDS_DATA } from './__fixtures__/parsedData.js';
+import { API_FIELDS_DATA } from '../__fixtures__/parsedData.js';
 
 const hydraSchemaAnalyzer = schemaAnalyzer();
 const dataProvider: ApiPlatformAdminDataProvider = {
