+++ src/field/FieldGuesser.tsx
@@ -28,7 +28,7 @@ import type {
 } from 'react-admin';
 import type { Field, Resource } from '@api-platform/api-doc-parser';
 
-import Introspecter from './Introspecter.js';
+import Introspecter from '../introspection/Introspecter.js';
 import EnumField from './EnumField.js';
 import type {
   EnumFieldProps,
@@ -36,7 +36,7 @@ import type {
   FieldProps,
   IntrospectedFieldGuesserProps,
   SchemaAnalyzer,
-} from './types.js';
+} from '../types.js';
 
 const isFieldSortable = (field: Field, schema: Resource) =>
   !!schema.parameters &&
