+++ src/index.ts
@@ -1,16 +1,16 @@
-import AdminGuesser from './AdminGuesser.js';
-import CreateGuesser from './CreateGuesser.js';
-import EditGuesser from './EditGuesser.js';
-import FieldGuesser from './FieldGuesser.js';
-import InputGuesser from './InputGuesser.js';
-import Introspecter from './Introspecter.js';
-import ListGuesser from './ListGuesser.js';
-import ResourceGuesser from './ResourceGuesser.js';
-import SchemaAnalyzerContext from './SchemaAnalyzerContext.js';
-import ShowGuesser from './ShowGuesser.js';
-import useIntrospect from './useIntrospect.js';
-import useIntrospection from './useIntrospection.js';
-import useMercureSubscription from './useMercureSubscription.js';
+import AdminGuesser from './core/AdminGuesser.js';
+import CreateGuesser from './create/CreateGuesser.js';
+import EditGuesser from './edit/EditGuesser.js';
+import FieldGuesser from './field/FieldGuesser.js';
+import InputGuesser from './input/InputGuesser.js';
+import Introspecter from './introspection/Introspecter.js';
+import ListGuesser from './list/ListGuesser.js';
+import ResourceGuesser from './core/ResourceGuesser.js';
+import SchemaAnalyzerContext from './introspection/SchemaAnalyzerContext.js';
+import ShowGuesser from './show/ShowGuesser.js';
+import useIntrospect from './introspection/useIntrospect.js';
+import useIntrospection from './introspection/useIntrospection.js';
+import useMercureSubscription from './mercure/useMercureSubscription.js';
 import useOnSubmit from './useOnSubmit.js';
 
 export {
