+++ src/introspection/SchemaAnalyzerContext.ts
@@ -1,7 +1,7 @@
 import {
   /* tree-shaking no-side-effects-when-called */ createContext,
 } from 'react';
-import type { SchemaAnalyzer } from './types.js';
+import type { SchemaAnalyzer } from '../types.js';
 
 const SchemaAnalyzerContext = createContext<SchemaAnalyzer | null>(null);
