+++ src/introspection/getRoutesAndResourcesFromNodes.tsx
