+++ src/openapi/OpenApiAdmin.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import dataProviderFactory from './dataProvider.js';
 import { restDataProvider } from '../dataProvider/index.js';
 import /* tree-shaking no-side-effects-when-called */ schemaAnalyzer from './schemaAnalyzer.js';
-import AdminGuesser from '../AdminGuesser.js';
-import type { AdminGuesserProps } from '../AdminGuesser.js';
+import AdminGuesser from '../core/AdminGuesser.js';
+import type { AdminGuesserProps } from '../core/AdminGuesser.js';
 import type { MercureOptions } from '../types.js';
 
 type AdminGuesserPartialProps = Omit<
